Generalize /writing/ into a topic-browsable home for all current writing
The section was scoped exclusively to Postgres content per an old
TigerData-influenced planning doc, leaving no home on the site for
general writing. `part` stays reserved for the curated 4-step Postgres
arc; `topics` becomes the universal tag mechanism for everything else,
with new /writing/tags/ browsing mirroring the existing legacy-archive
/tags/ pattern.
