+++ src/lib/posts.ts
@@ -1,38 +1,20 @@
 import type { CollectionEntry } from 'astro:content';
+import { collectTaxonomy as collectTaxonomyGeneric, entriesForTerm, slugify } from './taxonomy';
 
-/** Hugo-compatible urlize: lowercase, drop apostrophes, non-alnum → hyphen. */
-export function slugify(s: string): string {
-  return s
-    .toLowerCase()
-    .trim()
-    .replace(/['']/g, '')
-    .replace(/[^a-z0-9]+/g, '-')
-    .replace(/^-+|-+$/g, '');
-}
+export { slugify };
+export type { TaxoTerm } from './taxonomy';
 
 /** All legacy posts, newest first. */
 export function sortedPosts(posts: CollectionEntry<'posts'>[]) {
   return [...posts].sort((a, b) => b.data.date.getTime() - a.data.date.getTime());
 }
 
-export interface TaxoTerm { name: string; slug: string; count: number; }
-
 /** Build a slug→term registry for a taxonomy field, with post counts. */
 export function collectTaxonomy(
   posts: CollectionEntry<'posts'>[],
   field: 'categories' | 'tags',
-): TaxoTerm[] {
-  const bySlug = new Map<string, TaxoTerm>();
-  for (const p of posts) {
-    for (const name of p.data[field]) {
-      const slug = slugify(name);
-      if (!slug) continue;
-      const term = bySlug.get(slug);
-      if (term) term.count++;
-      else bySlug.set(slug, { name, slug, count: 1 });
-    }
-  }
-  return [...bySlug.values()].sort((a, b) => a.name.localeCompare(b.name));
+) {
+  return collectTaxonomyGeneric(posts, (p) => p.data[field]);
 }
 
 /** Posts whose taxonomy field contains a term matching the given slug. */
@@ -41,7 +23,7 @@ export function postsForTerm(
   field: 'categories' | 'tags',
   slug: string,
 ) {
-  return sortedPosts(posts.filter((p) => p.data[field].some((n) => slugify(n) === slug)));
+  return sortedPosts(entriesForTerm(posts, (p) => p.data[field], slug));
 }
 
 /** Group posts by year (descending) for the archive listing. */
